feat(core): SurfaceAp::read_surface now returns Surface instead of Vec<u8>
