+++ core/src/ap/mod.rs
@@ -1,6 +1,7 @@
 //! Collection of APs that corresponds with DF (Dedicated File) in the card
 
 pub mod jpki;
+pub mod support;
 pub mod surface;
 
 pub use self::jpki::JpkiAp;
